In order to de-couple commonly used code that is Linux specific from packages that might be used in non-Linux contexts, the sub-package, "connector", has been moved from the "endpoint" package to the "datapath" package.
